[2607-DEV-497] Stop bundling @tiptap into read-only guide/trip viewers
- library/[slug]/page.tsx (server component) now calls generateHTML
  itself and passes a plain HTML string to GuideBody, instead of
  passing JSONContent and letting the client component import
  @tiptap/html + starter-kit + extension-link + extension-image just
  to render static, admin-authored content.
- trips/[id]/components/shared.tsx: extracted TripDescription (the
  only tiptap consumer in that file) into its own component, loaded
  via next/dynamic(..., { ssr: false }) so @tiptap/* is code-split out
  of the initial bundle for every trip-detail view (Archived/Attendee/
  Available/PendingView all import shared.tsx). Threading server-
  rendered HTML through TripDetail's props chain (4 call sites) was
  more invasive than this ticket's scope for a bundle-size fix.

Closes #497
